% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earthquake}
\alias{earthquake}
\title{Earthquake data from Japan}
\format{
## `who`
A data frame with 62 rows and 2 columns:
\describe{
  \item{magnitude}{Magnitude of the earthquake}
  \item{direction}{Direction of the earthquake}
}
}
\source{
<https://is.gd/e6wKRk>
}
\usage{
earthquake
}
\description{
A tidy dataset of
}
\keyword{datasets}
