% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wildfire}
\alias{wildfire}
\title{Wildfire data from Portugal}
\format{
A data frame contains 102 rows and 2 columns:
\describe{
  \item{\emph{meanOrientation}}{Mean orientation of the fire}
  \item{\emph{burntArea}}{Mean burnt area of logs in ha}
}
}
\usage{
wildfire
}
\description{
A tidy dataset is about wildfires data, which occurred from
1985 to 2005 in Portugal.
}
\details{
There were 26870 observations of wildfire in Portugal between 1985
and 2005, then the orientation of the fire and the burnt area of logs were
aggregated in watersheds by the mean. For more details of the process of
cleaning of the data, the articles below give a better explanation.
}
\references{
\itemize{
   \item García-Portugués, Eduardo & Crujeiras, Rosa & Gonzãlez-Manteiga,
   Wenceslao. (2014). Central limit theorems for directional and linear random
   variables with applications. Statistica Sinica. 25. 10.5705/ss.2014.153.
   \item Ana M.G. Barros, José M.C. Pereira, Ulric J. Lund, Identifying
   geographical patterns of wildfire orientation: A watershed-based analysis,
   Forest Ecology and Management, Volume 264, 2012, Pages 98-107, ISSN 0378-1127,
   https://doi.org/10.1016/j.foreco.2011.09.027
}
}
\keyword{datasets}
